refactor: move sshnpd_doctor to a subdirectory and clarify --output option help text
