+++ packages/dart/noports_core/lib/sshnpd.dart
@@ -3,5 +3,5 @@ library noports_core_sshnpd;
 export 'src/sshnpd/sshnpd.dart';
 export 'src/sshnpd/sshnpd_params.dart';
 export 'src/sshnpd/sshnpd_config.dart';
-export 'src/sshnpd/sshnpd_doctor.dart';
+export 'src/sshnpd/doctor/sshnpd_doctor.dart';
 export 'src/common/types.dart';
